+++ src/components/string/strspn.c
@@ -20,5 +20,5 @@ size_t strspn(const char *s, const char *accept) {
 		++s_ptr;
 	}
 
-	return s_ptr - s;
+	return (size_t)(s_ptr - s);
 }
